Created separate component for the checkout cart - removed from Navbar component and migrated to child CartDrawer componenet
