+++ client/src/pages/Navbar/Navbar.jsx
@@ -1,28 +1,11 @@
-import AppBar from '@mui/material/AppBar';
-import Box from '@mui/material/Box';
-import Toolbar from '@mui/material/Toolbar';
-import Typography from '@mui/material/Typography';
-import { useState } from 'react';
+import {Typography, AppBar, Box, Toolbar, Container} from '@mui/material';
 import { NavLink } from 'react-router-dom';
-import ShoppingCartIcon from '@mui/icons-material/ShoppingCart';
-import { Badge } from '@mui/material';
-import { Container } from '@mui/material';
-import CheckoutMain from '../Checkout/CheckoutDrawer/CheckoutMain';
 import LogoDevIcon from '@mui/icons-material/LogoDev'; // Logo placeholder
 import MenuDrawer from './_tests_/MenuDrawer'; 
 import TopNavbar from './_tests_/TopNavbar';
+import CartDrawer from './_tests_/CartDrawer';
 
 export default function Navbar() {
-  const [showCart, setShowCart] = useState(false);
-
-
-  // Toggle CHECKOUT drawer
-  const toggleShowCart = () => {
-    setShowCart((prevShowCart) => !prevShowCart); //
-  };
-
-
-  // console.log('Auth Context: ', user);
 
   return (
     <Box sx={{ display: 'flex' }}>
@@ -34,12 +17,11 @@ export default function Navbar() {
         <Container maxWidth='xl'>
           <Toolbar disableGutters>
 
-
-            {/* MENU DRAWER - component*/}
+            {/* Menu drawer (mobile) */}
             <MenuDrawer />
 
 
-            {/* Navbars Logo & AppName */}
+            {/* Logo & Brand Name */}
             <Typography
               variant='h6'
               component='div'
@@ -49,32 +31,21 @@ export default function Navbar() {
               }}
             >
               <NavLink to='/' style={{ textDecoration: 'none', color: '#fff' }}>
+
                 <LogoDevIcon /> AppName
+
               </NavLink>
             </Typography>
 
-
-            {/* MAIN NAVBAR */}
+            {/* Main navbar */}
             <TopNavbar/>
 
-
-            {/* CHECKOUT DRAWER */}
-            <Box className='cart-icon' sx={{ ml: 3, mr: 4, cursor: 'pointer' }}>
-              <Badge
-                onClick={toggleShowCart}
-                badgeContent={1}
-                max={10}
-                color='error'
-              >
-                <ShoppingCartIcon />
-              </Badge>
-            </Box>
-            {showCart && <CheckoutMain />}
-
-
+            {/* Checkout drawer */}
+            <CartDrawer/>
 
           </Toolbar>
         </Container>
+
       </AppBar>
     </Box>
   );
